requirements.txt
gradio==4.0.2
hurry.filesize==0.9